add CoinManagerApduCommandsTest

+++ tonnfcclientandroid/src/main/java/com/tonnfccard/smartcard/ApduHelper.java
@@ -15,6 +15,10 @@ import static com.tonnfccard.smartcard.CommonConstants.SELECT_TON_WALLET_APPLET_
 import static com.tonnfccard.smartcard.TonWalletAppletApduCommands.WALLET_APPLET_CLA;
 import static com.tonnfccard.smartcard.TonWalletAppletApduCommands.getTonWalletAppletApduCommandName;
 
+/**
+ *  This utility class is to get the name of any APDU command.
+ */
+
 @RestrictTo(RestrictTo.Scope.LIBRARY)
 public class ApduHelper {
 
@@ -31,6 +35,7 @@ public class ApduHelper {
 
   private ApduHelper() {}
 
+  //Get the name of any APDU command, return null if command is not recognized
   public String getApduCommandName(CAPDU commandAPDU) {
     if (commandAPDU == null) throw new IllegalArgumentException(ERROR_MSG_CAPDU_IS_NULL);
     if (isSelectAPDU(commandAPDU)) {
@@ -40,12 +45,12 @@ public class ApduHelper {
     }
     if (commandAPDU.getCla() == WALLET_APPLET_CLA && getTonWalletAppletApduCommandName(commandAPDU.getIns()) != null)
       return getTonWalletAppletApduCommandName(commandAPDU.getIns());
-    if (commandAPDU.getCla() == (byte) COIN_MANAGER_CLA)
+    if (commandAPDU.getCla() == COIN_MANAGER_CLA)
       return getCoinManagerApduCommandName(BYTE_ARRAY_HELPER.hex(commandAPDU.getData()));
     return null;
   }
 
   private boolean isSelectAPDU(CAPDU commandAPDU) {
-    return commandAPDU.getCla() == SELECT_CLA && commandAPDU.getIns() == (byte) SELECT_INS &&  commandAPDU.getP1() == (byte) SELECT_P1 &&  commandAPDU.getP2() == (byte) SELECT_P2;
+    return commandAPDU.getCla() == SELECT_CLA && commandAPDU.getIns() == SELECT_INS &&  commandAPDU.getP1() == SELECT_P1 &&  commandAPDU.getP2() ==  SELECT_P2;
   }
 }
